--- src/about/index.html
@@ -1,19 +1,0 @@
-<!doctype html>
-<html lang="ja">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>About</title>
-  <link rel="stylesheet" href="/assets/style/style.scss" />
-  <script src="/assets/js/script.js" type="module"></script>
-</head>
-
-<body>
-  <p>Aboutページ</p>
-  <a href="/">トップへ戻る</a>
-  <div class="bg"></div>
-  <img src="/images/static.png" alt="" width="300" height="300" />
-</body>
-
-</html>
